introduce windows fast-forward workflow
The purpose of the fast-forwarding workflow is to aid in branch fast-forwarding
for the WMCO registry to avoid having to backport PRs from the release
branch to the community branch.

This workflow was adapted from
https://github.com/openshift/release/tree/0ef352953d988c2afcc5bf437ac2158ad603e858/ci-operator/step-registry/ocm/ci/fastforward
and changed to fit the needs of the WMCO including the use of the new
openshift-winc-bot to handle the fast-forward.

After adding the script, ran:
'sudo make update WHAT=openshift/windows-machine-config-operator
CONTAINER_ENGINE=podman'

Signed-off-by: Alina Ryan <aliryan@redhat.com>